feat: improved validation error messages
