+++ lint-staged.config.js
@@ -1,7 +1,4 @@
 module.exports = {
-  ignore: ['package-lock.json', 'CHANGELOG.md'],
-  linters: {
-    '*.js': ['prettier --write', 'eslint --fix', 'git add'],
-    '*.{json,md,yml,css}': ['prettier --write', 'git add'],
-  },
+  '*.js': ['prettier --write', 'eslint --fix', 'git add'],
+  '*.{json,md,yml,css}': ['prettier --write', 'git add'],
 };
